requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.61.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
ollama==0.5.3
packaging==25.0
pandas==3.0.0
pillow==12.1.0
prompt_toolkit==3.0.52
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
questionary==2.1.1
regex==2025.9.1
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.3.0
